+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added: `repo.saveAndSync` command (Ctrl + Shift + S)
 - Changed: `(New)` option moved to the bottom of omnimenu
+- Changed: Diff status display
 
 # v0.0.1